refactor: extract benchmark review lifecycle cli commands
