+++ loopx/cli_commands/__init__.py
@@ -16,6 +16,10 @@ from .benchmark_boundary import (
     handle_benchmark_boundary_command,
     register_benchmark_boundary_commands,
 )
+from .benchmark_review_lifecycle import (
+    handle_benchmark_review_lifecycle_command,
+    register_benchmark_review_lifecycle_commands,
+)
 from .doctor import handle_doctor_command, register_doctor_command
 from .dreaming import handle_dreaming_command, register_dreaming_commands
 from .history import handle_history_command, register_history_command
@@ -58,6 +62,7 @@ from .todo import handle_todo_command, register_todo_command
 __all__ = [
     "handle_agents_last_exam_command",
     "handle_benchmark_boundary_command",
+    "handle_benchmark_review_lifecycle_command",
     "handle_check_command",
     "handle_codex_cli_bounded_visible_pilot_adapter_command",
     "handle_codex_cli_bootstrap_message_command",
@@ -89,6 +94,7 @@ __all__ = [
     "handle_todo_command",
     "register_agents_last_exam_commands",
     "register_benchmark_boundary_commands",
+    "register_benchmark_review_lifecycle_commands",
     "register_doctor_command",
     "register_dreaming_commands",
     "register_history_command",
